@@ -64,7 +64,7 @@ describe Api::GraphqlController, type: :request do
 
     context 'with order not in submitted state' do
       before do
-        order.update_attributes! state: Order::PENDING
+        order.update! state: Order::PENDING
       end
       it 'returns error' do
         response = client.execute(mutation, approve_order_input)
@@ -77,7 +77,7 @@ describe Api::GraphqlController, type: :request do
     context 'with proper permission' do
       before do
         Fabricate(:transaction, order: order, external_id: 'pi_1', external_type: Transaction::PAYMENT_INTENT)
-        order.update_attributes! state: Order::SUBMITTED, external_charge_id: 'pi_1'
+        order.update! state: Order::SUBMITTED, external_charge_id: 'pi_1'
         prepare_payment_intent_capture_success(amount: 20_00)
       end
       it 'approves the order' do
