@@ -1,10 +1,5 @@
 require 'sidekiq/web'
 Rails.application.routes.draw do
-  ActiveAdmin.routes(self)
-  mount ArtsyAuth::Engine => '/'
-  constraints ->(req) { req.session[:access_token].present? && ArtsyAuthToken.new(req.session[:access_token]).admin? } do
-    mount Sidekiq::Web, at: '/admin/sidekiq'
-  end
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
   namespace :api do
     post '/graphql', to: 'graphql#execute'
@@ -12,6 +7,11 @@ Rails.application.routes.draw do
     post '/webhooks/stripe', to: 'webhooks#stripe'
   end
   resources :admin_notes
+  ActiveAdmin.routes(self)
+  mount ArtsyAuth::Engine => '/'
+  constraints ->(req) { req.session[:access_token].present? && ArtsyAuthToken.new(req.session[:access_token]).admin? } do
+    mount Sidekiq::Web, at: '/admin/sidekiq'
+  end
 
   root to: redirect('/admin')
 end
