@@ -8,7 +8,7 @@ RSpec.shared_context 'include stripe helper' do
       payment_intent = double(id: id, payment_method: payment_method, capture_method: capture ? 'automatic' : 'manual', amount: amount, status: status, client_secret: client_secret)
       mock_payment_intent_call(:create, payment_intent)
     when 'requires_payment_method'
-      error = Stripe::CardError.new(charge_error[:message], charge_error[:decline_code], charge_error[:code])
+      error = Stripe::CardError.new(charge_error[:message], decline_code: charge_error[:decline_code], code: charge_error[:code])
       allow(error).to receive(:json_body).and_return(error: { payment_intent: basic_payment_intent(status: status, capture: capture, amount: amount, code: charge_error[:code], decline_code: charge_error[:decline_code]) })
       allow(Stripe::PaymentIntent).to receive(:create).and_raise(error)
     when 'testmode_charges_only'
@@ -25,7 +25,7 @@ RSpec.shared_context 'include stripe helper' do
 
   def prepare_payment_intent_confirm_failure(id: 'pi_1', charge_error:, payment_method: 'cc_1', amount: 20_00, status: 'requires_confirmation')
     payment_intent = double(id: id, payment_method: payment_method, amount: amount, capture_method: 'manual', status: status, transfer_data: double(destination: 'ma_1'), last_payment_error: double(charge_error))
-    error = Stripe::CardError.new(charge_error[:message], charge_error[:decline_code], charge_error[:code])
+    error = Stripe::CardError.new(charge_error[:message], decline_code: charge_error[:decline_code], code: charge_error[:code])
     allow(payment_intent).to receive(:confirm).and_raise(error)
     allow(error).to receive(:json_body).and_return(error: { payment_intent: basic_payment_intent(status: 'requires_payment_method', capture: true, amount: amount, code: charge_error[:code], decline_code: charge_error[:decline_code]) })
     mock_payment_intent_call(:retrieve, payment_intent)
@@ -56,7 +56,7 @@ RSpec.shared_context 'include stripe helper' do
 
   def prepare_payment_intent_capture_failure(charge_error:, payment_method: 'cc_1', amount: 20_00)
     payment_intent = double(id: 'pi_1', payment_method: payment_method, amount: amount, capture_method: 'manual', status: 'requires_capture', transfer_data: double(destination: 'ma_1'), last_payment_error: double(charge_error))
-    error = Stripe::CardError.new(charge_error[:message], charge_error[:decline_code], charge_error[:code])
+    error = Stripe::CardError.new(charge_error[:message], decline_code: charge_error[:decline_code], code: charge_error[:code])
     allow(payment_intent).to receive(:capture).and_raise(error)
     allow(error).to receive(:json_body).and_return(error: { payment_intent: basic_payment_intent(status: 'requires_payment_method', capture: true, amount: amount, code: charge_error[:code], decline_code: charge_error[:decline_code]) })
     mock_payment_intent_call(:retrieve, payment_intent)
@@ -71,7 +71,7 @@ RSpec.shared_context 'include stripe helper' do
 
   def prepare_payment_intent_cancel_failure(charge_error:, payment_method: 'cc_1', amount: 20_00)
     payment_intent = double(id: 'pi_1', payment_method: payment_method, amount: amount, capture_method: 'manual', status: 'requires_capture', transfer_data: double(destination: 'ma_1'), last_payment_error: double(charge_error))
-    error = Stripe::CardError.new(charge_error[:message], charge_error[:decline_code], charge_error[:code])
+    error = Stripe::CardError.new(charge_error[:message], decline_code: charge_error[:decline_code], code: charge_error[:code])
     allow(payment_intent).to receive(:cancel).and_raise(error)
     allow(error).to receive(:json_body).and_return(error: { payment_intent: basic_payment_intent(status: 'requires_payment_method', capture: true, amount: amount, code: charge_error[:code], decline_code: charge_error[:decline_code]) })
     mock_payment_intent_call(:retrieve, payment_intent)
@@ -109,7 +109,7 @@ RSpec.shared_context 'include stripe helper' do
   end
 
   def prepare_setup_intent_create_failure(charge_error:)
-    error = Stripe::CardError.new(charge_error[:message], charge_error[:decline_code], charge_error[:code])
+    error = Stripe::CardError.new(charge_error[:message], decline_code: charge_error[:decline_code], code: charge_error[:code])
     allow(error).to receive(:json_body).and_return(error: { setup_intent: basic_setup_intent(code: charge_error[:code], decline_code: charge_error[:decline_code]) })
     allow(Stripe::SetupIntent).to receive(:create).and_raise(error)
   end
